# Conflicts:
#	api/src/window.ts
#	cli/core/Cargo.lock
#	cli/core/schema.json
#	cli/tauri-bundler/Cargo.lock
#	cli/tauri-bundler/Cargo.toml
#	cli/tauri-bundler/src/bundle/settings.rs
#	cli/tauri-bundler/src/bundle/tauri_config.rs

+++ cli/core/src/build.rs
@@ -1,7 +1,4 @@
-use tauri_bundler::{
-  build_project,
-  bundle::{bundle_project, PackageType, SettingsBuilder},
-};
+use tauri_bundler::bundle::{bundle_project, PackageType, SettingsBuilder};
 
 use crate::helpers::{
   app_paths::{app_dir, tauri_dir},
@@ -10,13 +7,10 @@ use crate::helpers::{
   manifest::rewrite_manifest,
   Logger, TauriScript,
 };
-use std::{
-  env::{set_current_dir, set_var},
-  fs::File,
-  io::Write,
-  path::PathBuf,
-  process::Command,
-};
+
+use std::{env::set_current_dir, fs::File, io::Write, path::PathBuf, process::Command};
+
+mod rust;
 
 #[derive(Default)]
 pub struct Build {
@@ -54,43 +48,10 @@ impl Build {
   pub fn run(self) -> crate::Result<()> {
     let logger = Logger::new("tauri:build");
     let config = get_config(self.config.as_deref())?;
-    let config_guard = config.lock().unwrap();
-    let config_ = config_guard.as_ref().unwrap();
-
-    let mut settings_builder = SettingsBuilder::new().features(vec!["embedded-server".to_string()]);
-    if !self.debug {
-      settings_builder = settings_builder.release();
-    }
-    if self.verbose {
-      settings_builder = settings_builder.verbose();
-    }
-    if let Some(names) = self.targets {
-      let mut types = vec![];
-      for name in names {
-        if name == "none" {
-          break;
-        }
-        match PackageType::from_short_name(&name) {
-          Some(package_type) => {
-            types.push(package_type);
-          }
-          None => {
-            return Err(anyhow::anyhow!(format!(
-              "Unsupported bundle format: {}",
-              name
-            )));
-          }
-        }
-      }
-      settings_builder = settings_builder.package_types(types);
-    }
 
     let tauri_path = tauri_dir();
     set_current_dir(&tauri_path)?;
-    set_var("TAURI_DIR", &tauri_path);
-    set_var("TAURI_DIST_DIR", tauri_path.join(&config_.build.dist_dir));
 
-    drop(config_guard);
     rewrite_manifest(config.clone())?;
 
     let config_guard = config.lock().unwrap();
@@ -100,11 +61,16 @@ impl Build {
     let tauri_script = TauriScript::new()
       .global_tauri(config_.build.with_global_tauri)
       .get();
-    let tauri_script_path = PathBuf::from(&config_.build.dist_dir).join("__tauri.js");
+    let web_asset_path = PathBuf::from(&config_.build.dist_dir);
+    if !web_asset_path.exists() {
+      return Err(anyhow::anyhow!(
+        "Unable to find your web assets, did you forget to build your web app? Your distDir is set to \"{:?}\".",
+        web_asset_path
+      ));
+    }
+    let tauri_script_path = web_asset_path.join("__tauri.js");
     let mut tauri_script_file = File::create(tauri_script_path)?;
     tauri_script_file.write_all(tauri_script.as_bytes())?;
-
-    let settings = settings_builder.build()?;
 
     if let Some(before_build) = &config_.build.before_build_command {
       let mut cmd: Option<&str> = None;
@@ -130,10 +96,43 @@ impl Build {
       }
     }
 
-    build_project(&settings)?;
+    rust::build_project(self.debug)?;
+
     if config_.tauri.bundle.active {
+      let bundler_settings = rust::get_bundler_settings(&config_, self.debug)?;
+      let mut settings_builder = SettingsBuilder::new()
+        .package_settings(bundler_settings.package_settings)
+        .bundle_settings(bundler_settings.bundle_settings)
+        .binaries(bundler_settings.binaries)
+        .project_out_directory(bundler_settings.out_dir);
+
+      if self.verbose {
+        settings_builder = settings_builder.verbose();
+      }
+      if let Some(names) = self.targets {
+        let mut types = vec![];
+        for name in names {
+          if name == "none" {
+            break;
+          }
+          match PackageType::from_short_name(&name) {
+            Some(package_type) => {
+              types.push(package_type);
+            }
+            None => {
+              return Err(anyhow::anyhow!(format!(
+                "Unsupported bundle format: {}",
+                name
+              )));
+            }
+          }
+        }
+        settings_builder = settings_builder.package_types(types);
+      }
+      let settings = settings_builder.build()?;
       bundle_project(settings)?;
     }
+
     Ok(())
   }
 }
